src/main/java/com/tide/ami/service/impl/UserServiceImpl.java
package com.tide.ami.service.impl;

import javax.transaction.Transactional;

import org.springframework.stereotype.Service;
import org.springframework.util.Assert;

import com.tide.ami.commands.RegistrationCommand;
import com.tide.ami.event.DomainEventPublisher;
import com.tide.ami.event.UserRegisteredEvent;
import com.tide.ami.exception.RegistrationException;
import com.tide.ami.mail.MailManager;
import com.tide.ami.mail.MessageVariable;
import com.tide.ami.model.RegistrationManagement;
import com.tide.ami.model.User;
import com.tide.ami.service.UserService;

@Service
@Transactional
public class UserServiceImpl implements UserService {

  private RegistrationManagement registrationManagement;
  private DomainEventPublisher domainEventPublisher;
  private MailManager mailManager;

  public UserServiceImpl(RegistrationManagement registrationManagement,
                         DomainEventPublisher domainEventPublisher,
                         MailManager mailManager) {
    this.registrationManagement = registrationManagement;
    this.domainEventPublisher = domainEventPublisher;
    this.mailManager = mailManager;
  }

  @Override
  public void register(RegistrationCommand command) throws RegistrationException {
    Assert.notNull(command, "Parameter `command` must not be null");
    User newUser = registrationManagement.register(
      command.getUsername(),
      command.getEmailAddress(),
      command.getPassword());

    sendWelcomeMessage(newUser);
    domainEventPublisher.publish(new UserRegisteredEvent(newUser));
  }

  private void sendWelcomeMessage(User user) {
    mailManager.send(
      user.getEmailAddress(),
      "Welcome to TaskAgile",
      "welcome.ftl",
      MessageVariable.from("user", user)
    );
  }
}
